move `gallery/` to `site/`
